+++ FrontEndScripts/Filters/Markdown.ts
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright © Aleksey Nemiro, 2015. All rights reserved.
+ * Copyright © Aleksey Nemiro, 2016. All rights reserved.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.