+++ SmallServerAdmin/servers/default.php
@@ -1,0 +1,16 @@
+<?php
+
+// Specify in this file the parameters of connection to the server.
+// WARNING: DO NOT DECLARE $config
+
+// server config
+$config['server_name'] = 'Main Server';
+$config['server_description'] = 'My most main server.';
+# $config['server_disabled'] = TRUE;
+
+// ssh
+$config['ssh_host'] = '192.168.56.139';
+$config['ssh_port'] = '22';
+$config['ssh_user'] = 'ssh username';
+$config['ssh_password'] = 'ssh password here';
+$config['ssh_required_password'] = TRUE;